+++ src/chart_types/xy_chart/utils/dimensions.ts
@@ -19,8 +19,13 @@ export function computeChartDimensions(
   chartTheme: Theme,
   axisDimensions: Map<AxisId, AxisTicksDimensions>,
   axisSpecs: Map<AxisId, AxisSpec>,
-): Dimensions {
-  const { chartMargins, chartPaddings } = chartTheme;
+  showLegend: boolean,
+  legendPosition?: Position,
+): {
+  chartDimensions: Dimensions;
+  leftMargin: number;
+} {
+  const { chartMargins, chartPaddings, legend } = chartTheme;
   const { axisTitleStyle } = chartTheme.axes;
 
   const axisTitleHeight = axisTitleStyle.fontSize + axisTitleStyle.padding;
@@ -29,7 +34,8 @@ export function computeChartDimensions(
   let vRightAxisSpecWidth = 0;
   let hTopAxisSpecHeight = 0;
   let hBottomAxisSpecHeight = 0;
-
+  let horizontalEdgeLabelOverflow = 0;
+  let verticalEdgeLabelOverflow = 0;
   axisDimensions.forEach(({ maxLabelBboxWidth = 0, maxLabelBboxHeight = 0 }, id) => {
     const axisSpec = axisSpecs.get(id);
     if (!axisSpec || axisSpec.hide) {
@@ -37,55 +43,72 @@ export function computeChartDimensions(
     }
     const { position, tickSize, tickPadding, title } = axisSpec;
     const titleHeight = title !== undefined ? axisTitleHeight : 0;
+    const maxAxisHeight = maxLabelBboxHeight + tickSize + tickPadding + titleHeight;
+    const maxAxisWidth = maxLabelBboxWidth + tickSize + tickPadding + titleHeight;
     switch (position) {
       case Position.Top:
-        hTopAxisSpecHeight += maxLabelBboxHeight + tickSize + tickPadding + chartMargins.top + titleHeight;
+        hTopAxisSpecHeight += maxAxisHeight + chartMargins.top;
+        // find the max half label size to accomodate the left/right labels
+        horizontalEdgeLabelOverflow = Math.max(horizontalEdgeLabelOverflow, maxLabelBboxWidth / 2);
         break;
       case Position.Bottom:
-        hBottomAxisSpecHeight += maxLabelBboxHeight + tickSize + tickPadding + chartMargins.bottom + titleHeight;
+        hBottomAxisSpecHeight += maxAxisHeight + chartMargins.bottom;
+        // find the max half label size to accomodate the left/right labels
+        horizontalEdgeLabelOverflow = Math.max(horizontalEdgeLabelOverflow, maxLabelBboxWidth / 2);
         break;
       case Position.Left:
-        vLeftAxisSpecWidth += maxLabelBboxWidth + tickSize + tickPadding + chartMargins.left + titleHeight;
+        vLeftAxisSpecWidth += maxAxisWidth + chartMargins.left;
+        verticalEdgeLabelOverflow = Math.max(verticalEdgeLabelOverflow, maxLabelBboxHeight / 2);
         break;
       case Position.Right:
-        vRightAxisSpecWidth += maxLabelBboxWidth + tickSize + tickPadding + chartMargins.right + titleHeight;
+        vRightAxisSpecWidth += maxAxisWidth + chartMargins.right;
+        verticalEdgeLabelOverflow = Math.max(verticalEdgeLabelOverflow, maxLabelBboxHeight / 2);
         break;
     }
   });
-  // const hMargins = chartMargins.left + chartMargins.right;
-  const chartWidth = parentDimensions.width - vLeftAxisSpecWidth - vRightAxisSpecWidth;
-  const chartHeight = parentDimensions.height - hTopAxisSpecHeight - hBottomAxisSpecHeight;
+  const chartLeftAxisMaxWidth = Math.max(vLeftAxisSpecWidth, horizontalEdgeLabelOverflow + chartMargins.left);
+  const chartRightAxisMaxWidth = Math.max(vRightAxisSpecWidth, horizontalEdgeLabelOverflow + chartMargins.right);
+  const chartTopAxisMaxHeight = Math.max(hTopAxisSpecHeight, verticalEdgeLabelOverflow + chartMargins.top);
+  const chartBottomAxisMaxHeight = Math.max(hBottomAxisSpecHeight, verticalEdgeLabelOverflow + chartMargins.bottom);
+
+  const chartWidth = parentDimensions.width - chartLeftAxisMaxWidth - chartRightAxisMaxWidth;
+  const chartHeight = parentDimensions.height - chartTopAxisMaxHeight - chartBottomAxisMaxHeight;
+
   let vMargin = 0;
-  if (hTopAxisSpecHeight === 0) {
-    vMargin += chartMargins.top;
-  }
-  if (hBottomAxisSpecHeight === 0) {
-    vMargin += chartMargins.bottom;
-  }
   let hMargin = 0;
-  if (vLeftAxisSpecWidth === 0) {
-    hMargin += chartMargins.left;
-  }
-  if (vRightAxisSpecWidth === 0) {
-    hMargin += chartMargins.right;
+
+  // add space for legend
+  let legendTopMargin = 0;
+  let legendLeftMargin = 0;
+  if (showLegend && legendPosition) {
+    switch (legendPosition) {
+      case Position.Right:
+        hMargin += legend.verticalWidth;
+        break;
+      case Position.Left:
+        hMargin += legend.verticalWidth;
+        legendLeftMargin = legend.verticalWidth;
+        break;
+      case Position.Top:
+        vMargin += legend.horizontalHeight;
+        legendTopMargin = legend.horizontalHeight;
+        break;
+      case Position.Bottom:
+        vMargin += legend.horizontalHeight;
+        break;
+    }
   }
 
-  let top = 0;
-  let left = 0;
-  if (hTopAxisSpecHeight === 0) {
-    top = chartMargins.top + chartPaddings.top;
-  } else {
-    top = hTopAxisSpecHeight + chartPaddings.top;
-  }
-  if (vLeftAxisSpecWidth === 0) {
-    left = chartMargins.left + chartPaddings.left;
-  } else {
-    left = vLeftAxisSpecWidth + chartPaddings.left;
-  }
+  let top = chartTopAxisMaxHeight + chartPaddings.top + legendTopMargin;
+  let left = chartLeftAxisMaxWidth + chartPaddings.left + legendLeftMargin;
+
   return {
-    top,
-    left,
-    width: chartWidth - hMargin - chartPaddings.left - chartPaddings.right,
-    height: chartHeight - vMargin - chartPaddings.top - chartPaddings.bottom,
+    leftMargin: chartLeftAxisMaxWidth - vLeftAxisSpecWidth,
+    chartDimensions: {
+      top,
+      left,
+      width: chartWidth - hMargin - chartPaddings.left - chartPaddings.right,
+      height: chartHeight - vMargin - chartPaddings.top - chartPaddings.bottom,
+    },
   };
 }